requirements.txt
# Python dependencies for Raspberry Pi Voice Assistant
# Install with: pip install -r requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alibabacloud-bailian20230601==1.6.1
alibabacloud-bailian20231229==2.8.1
alibabacloud-credentials==1.0.5
alibabacloud-credentials-api==1.0.0
alibabacloud-tea==0.4.3
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.4
alibabacloud_tea_openapi==0.4.3
alibabacloud_tea_util==0.3.14
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
broadscope-bailian==1.3.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
crcmod==1.7
cryptography==44.0.3
darabonba-core==1.0.5
dashscope==1.25.11
decorator==5.2.1
editdistance==0.8.1
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
funasr==1.3.1
gradio==6.4.0
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface_hub==1.3.2
hydra-core==1.3.2
idna==3.11
jaconv==0.4.1
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.3
kaldiio==2.18.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.34.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.11.5
oss2==2.19.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.4
PyAudio==0.2.14
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pynndescent==0.6.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytorch-wpe==0.0.1
pyttsx3==2.99
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
RPi.GPIO==0.7.1
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.50.0
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.7.0
torch-complex==0.4.4
torchaudio==2.7.0
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.11
urllib3==2.6.3
uvicorn==0.40.0
vosk==0.3.45
websocket-client==1.9.0
websockets==11.0.3
yarl==1.22.0

# OLED显示相关依赖
luma.oled==3.12.0
luma.core==2.3.1
